Move `AnalyzerNetworkBase` and `ReverseAnalyzerBase` into separate files
update imports accordingly and remove "comment walls".
