+++ src/innvestigate/analyzer/pattern_based.py
@@ -1,7 +1,7 @@
-# Get Python six functionality:
-from __future__ import absolute_import, division, print_function, unicode_literals
+from __future__ import annotations
 
 import warnings
+from typing import Dict
 
 import keras
 import keras.activations
@@ -12,28 +12,19 @@ import keras.layers.pooling
 import keras.models
 import numpy as np
 
-from .. import layers as ilayers
-from .. import tools as itools
-from .. import utils
-from ..utils import keras as kutils
-from ..utils.keras import checks as kchecks
-from ..utils.keras import graph as kgraph
-from . import base
-
-###############################################################################
-###############################################################################
-###############################################################################
-
+import innvestigate.layers as ilayers
+import innvestigate.tools as itools
+import innvestigate.utils as iutils
+import innvestigate.utils.keras as kutils
+import innvestigate.utils.keras.checks as kchecks
+import innvestigate.utils.keras.graph as kgraph
+from innvestigate.analyzer.base import OneEpochTrainerMixin
+from innvestigate.analyzer.reverse_base import ReverseAnalyzerBase
 
 __all__ = [
     "PatternNet",
     "PatternAttribution",
 ]
-
-
-###############################################################################
-###############################################################################
-###############################################################################
 
 
 SUPPORTED_LAYER_PATTERNNET = (
@@ -109,13 +100,13 @@ class PatternNetReverseKernelLayer(kgraph.ReverseMappingBase):
             tmp = reversed_Ys
         else:
             # if linear activation this behaves strange
-            tmp = utils.to_list(grad_act(act_Xs + act_Ys + reversed_Ys))
+            tmp = iutils.to_list(grad_act(act_Xs + act_Ys + reversed_Ys))
 
         # Second step: propagate through the pattern layer.
         return grad_pattern(Xs + pattern_Ys + tmp)
 
 
-class PatternNet(base.OneEpochTrainerMixin, base.ReverseAnalyzerBase):
+class PatternNet(OneEpochTrainerMixin, ReverseAnalyzerBase):
     """PatternNet analyzer.
 
     Applies the "PatternNet" algorithm to analyze the model's predictions.
