+++ src/innvestigate/analyzer/relevance_based/utils.py
@@ -1,17 +1,6 @@
-# Get Python six functionality:
-from __future__ import absolute_import, division, print_function, unicode_literals
-
-###############################################################################
-###############################################################################
-###############################################################################
-
+from __future__ import annotations
 
 __all__ = ["assert_lrp_epsilon_param", "assert_infer_lrp_alpha_beta_param"]
-
-
-###############################################################################
-###############################################################################
-###############################################################################
 
 
 def assert_lrp_epsilon_param(epsilon, caller):
@@ -117,6 +106,4 @@ def assert_infer_lrp_alpha_beta_param(alpha, beta, caller):
     return alpha, beta
 
 
-###############################################################################
-###############################################################################
 ###############################################################################
